chore: refactor tx-builder
