+++ src/web/index.ts
@@ -2,4 +2,4 @@ export * from '../common/bcs';
 export * from '../common/utils';
 export * from '../common/types';
 export * from '../common/execute';
-export * from './tx-builder';
+export * from '../common/tx-builder';
